refactor: extract canvas keyboard handler into useCanvasKeyboard hook
Move handleKeyDown (~510 lines) from SketchCanvas.tsx into new
useCanvasKeyboard.ts hook. Behavior-preserving verbatim move following
the established ctx/return convention from prior workflow hooks.

- New file: useCanvasKeyboard.ts (707 lines) exports CanvasKeyboardCtx
  interface and useCanvasKeyboard function returning { handleKeyDown }
- SketchCanvas shrinks from 5577 to 5134 lines (-443)
- All input-level onKeyDown handlers and makeEditInputKeyDown remain
  untouched in the shell
- Canvas binding: onKeyDown={keyboard.handleKeyDown}
